requirements.txt
pyserial
supabase
python-dotenv
pytz